+++ tests/test_signoff_flow.py
@@ -372,6 +372,73 @@ class UserGroupsTest(PostgresWebTest, FormattedErrorMixin, unittest.TestCase):
                             headers=self.reviewer_headers)
 
 
+class SpecificUserGroupsTest(PostgresWebTest, FormattedErrorMixin, unittest.TestCase):
+    def get_app_settings(self, extras=None):
+        settings = super(SpecificUserGroupsTest, self).get_app_settings(extras)
+
+        self.source_collection1 = "/buckets/alice/collections/cid1"
+        self.source_collection2 = "/buckets/alice/collections/cid2"
+
+        settings['kinto.signer.resources'] = "%s;%s %s;%s" % (
+            self.source_collection1,
+            self.source_collection1.replace("alice", "destination"),
+            self.source_collection2,
+            self.source_collection2.replace("alice", "destination"))
+
+        settings['signer.group_check_enabled'] = 'false'
+        settings['signer.alice_cid1_group_check_enabled'] = 'true'
+        settings['signer.alice_cid1_editors_group'] = 'editeurs'
+        settings['signer.alice_cid1_reviewers_group'] = 'revoyeurs'
+        return settings
+
+    def setUp(self):
+        super(SpecificUserGroupsTest, self).setUp()
+
+        self.app.put_json(self.source_collection1, headers=self.headers)
+        self.app.put_json(self.source_collection2, headers=self.headers)
+
+        self.someone_headers = get_user_headers('sam:wan')
+
+        self.editor_headers = get_user_headers('emo:billier')
+        resp = self.app.get("/", headers=self.editor_headers)
+        self.editor = resp.json["user"]["id"]
+
+        self.app.put_json("/buckets/alice/groups/editeurs",
+                          {"data": {"members": [self.editor]}},
+                          headers=self.headers)
+
+    def test_editors_can_ask_to_review_if_not_specificly_configured(self):
+        self.app.patch_json(self.source_collection2,
+                            {"data": {"status": "to-review"}},
+                            headers=self.someone_headers,
+                            status=200)
+
+    def test_only_specific_editors_can_ask_to_review(self):
+        resp = self.app.patch_json(self.source_collection1,
+                                   {"data": {"status": "to-review"}},
+                                   headers=self.someone_headers,
+                                   status=403)
+        self.assertFormattedError(response=resp,
+                                  code=403,
+                                  errno=ERRORS.FORBIDDEN,
+                                  error="Forbidden",
+                                  message="Not in editeurs group")
+
+    def test_only_reviewers_can_ask_to_sign(self):
+        self.app.patch_json(self.source_collection1,
+                            {"data": {"status": "to-review"}},
+                            headers=self.editor_headers)
+        resp = self.app.patch_json(self.source_collection1,
+                                   {"data": {"status": "to-sign"}},
+                                   headers=self.editor_headers,
+                                   status=403)
+        self.assertFormattedError(response=resp,
+                                  code=403,
+                                  errno=ERRORS.FORBIDDEN,
+                                  error="Forbidden",
+                                  message="Not in revoyeurs group")
+
+
 class PreviewCollectionTest(PostgresWebTest, unittest.TestCase):
     def get_app_settings(self, extras=None):
         settings = super(PreviewCollectionTest, self).get_app_settings(extras)